- DB geupdated mit front und back

+++ app/src/main/java/de/beacon/tom/viibenav_radiomapper/model/position/neighbor/DeviationToCoord.java
@@ -8,19 +8,19 @@ import de.beacon.tom.viibenav_radiomapper.model.Coordinate;
  */
 public class DeviationToCoord {
 
-    private double deviation;
+    private float deviation;
     private Coordinate coordinate;
 
-    public DeviationToCoord(double deviation, Coordinate coordinate) {
+    public DeviationToCoord(float deviation, Coordinate coordinate) {
         this.deviation = deviation;
         this.coordinate = coordinate;
     }
 
-    public double getdeviation() {
+    public float getdeviation() {
         return deviation;
     }
 
-    public void setdeviation(double deviation) {
+    public void setdeviation(float deviation) {
         this.deviation = deviation;
     }
